quants/requirements.txt
Pillow==9.5.0
PyYAML==6.0
torch==2.0.0
torchvision==0.15.1
pytorch-lightning~=1.6.5
timm~=0.6.7